requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
Pillow
python-multipart
google-generativeai
python-dotenv